meiduo_mall/requirements.txt
celery==5.1.0
configparser==5.0.2
Django==3.1.7
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-haystack==3.0
django-js-asset==1.2.2
django-redis==4.12.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-extensions==0.7.0
drf-haystack==1.8.10
elasticsearch==2.4.1
itsdangerous==1.1.0
mysqlclient==2.0.3
Pillow==8.2.0
py3Fdfs==2.2.0
PyJWT==1.7.1
PyMySQL==1.0.2
QQLoginTool==0.3.0
redis==3.5.3
requests==2.25.1
urllib3==1.26.3

